requirements.txt
transformers==4.28.1
diffusers==0.15.0
torch==2.0.0
torchvision==0.15.1
streamlit==1.21.0
pillow==9.5.0
pika==1.3.1
accelerate==0.18.0